:recycle: Refactor tree

+++ src/commands/dev.ts
@@ -1,10 +1,11 @@
 import fs from "fs"
-import path from 'path'
+import path from "path"
+import yaml from "js-yaml"
 import consola from "consola"
 
-import bs from 'browser-sync'
-import { Ploceus } from '../core'
-import { Command, flags } from '@oclif/command'
+import bs from "browser-sync"
+import { Ploceus, SiteConfig } from "../core"
+import { Command, flags } from "@oclif/command"
 
 export default class Dev extends Command {
     static description = 'Start development mode.'
@@ -26,6 +27,17 @@ export default class Dev extends Command {
 
         const sourcePath = path.resolve(args.path)
         const distPath = path.resolve(sourcePath, "dist")
+        const siteConfPath = path.join(sourcePath, "site.yaml")
+
+        if (!fs.existsSync(siteConfPath)) {
+            consola.error("site.yaml not found")
+            process.exit(1)
+        }
+
+        let siteConfig = yaml.load(
+            fs.readFileSync(siteConfPath).toString()
+        ) as SiteConfig
+
         fs.rmSync(distPath, { recursive: true, force: true })
         fs.mkdirSync(distPath, { recursive: true })
 
@@ -38,7 +50,7 @@ export default class Dev extends Command {
         })
 
         new Ploceus(sourcePath, {
-            dev: true, production: false
+            dev: true, production: false, siteConfig
         }).on("ready", () => {
             consola.ready(`development server started at http://localhost:${bsInstance.getOption("port")}`)
         }).on("error", error => {
